% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentAnalysis.R
\name{FindAllDEPs}
\alias{FindAllDEPs}
\title{Find All Differentially Expressed Programs (1-vs-all analysis)}
\usage{
FindAllDEPs(
  x,
  cell_type_col,
  nmf_name = "NMF",
  test = "wilcox",
  pval.type = "some",
  log.p = FALSE,
  min.prop = 0.1
)
}
\arguments{
\item{x}{A SingleCellExperiment object with NMF results}

\item{cell_type_col}{Character, column name in colData containing cell type labels}

\item{nmf_name}{Character, name of NMF result to use (default "NMF")}

\item{test}{Character, statistical test to use: "wilcox" (default), "t", "binom"}

\item{pval.type}{Character, p-value type: "some", "any", "all" (default "some")}

\item{log.p}{Logical, whether to log-transform p-values (default FALSE)}

\item{min.prop}{Numeric, minimum proportion of cells that must express program (default 0.1)}
}
\value{
List of DataFrames (one per cell type) with differential expression statistics
}
\description{
Uses scran::findMarkers to identify NMF programs that are differentially
expressed in each cell type compared to all other cell types. Treats each
NMF program as a "gene" and performs 1-vs-all comparisons.
}
\examples{
library(scuttle)
sce <- mockSCE(ngenes = 100, ncells = 50)
sce <- logNormCounts(sce)
sce <- runNMFscape(sce, k = 3, verbose = FALSE)
sce$celltype <- sample(c("A", "B"), ncol(sce), replace = TRUE)
deps <- FindAllDEPs(sce, cell_type_col = "celltype")
head(deps[[1]])

}
