% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getDiagonal}
\alias{getDiagonal}
\title{Extract diagonal scaling vector from NMF decomposition}
\usage{
getDiagonal(x, name = "NMF")
}
\arguments{
\item{x}{A SingleCellExperiment object with NMF results}

\item{name}{Character, name of the NMF result (default "NMF")}
}
\value{
Numeric vector of length k (diagonal scaling factors)
}
\description{
Returns the diagonal scaling vector d from the decomposition A = W * diag(d) * H.
}
\examples{
library(scuttle)
sce <- mockSCE(ngenes = 100, ncells = 50)
sce <- logNormCounts(sce)
sce <- runNMFscape(sce, k = 3, verbose = FALSE)
getDiagonal(sce)
}
