% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getModel}
\alias{getModel}
\title{Extract raw RcppML NMF model from SingleCellExperiment object}
\usage{
getModel(x, name = "NMF")
}
\arguments{
\item{x}{A SingleCellExperiment object with NMF results}

\item{name}{Character, name of the NMF result (default "NMF")}
}
\value{
An S4 object of class \code{nmf} from RcppML with slots
  \code{@w}, \code{@d}, \code{@h}, \code{@misc}
}
\description{
Returns the raw RcppML S4 nmf model object stored by \code{\link{runNMFscape}}.
Useful for passing to RcppML functions like \code{predict},
\code{refine}, \code{align}, etc.
}
\examples{
library(scuttle)
sce <- mockSCE(ngenes = 100, ncells = 50)
sce <- logNormCounts(sce)
sce <- runNMFscape(sce, k = 3, verbose = FALSE)
model <- getModel(sce)
dim(model@w)
}
