% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateNMF.R
\name{evaluateNMF}
\alias{evaluateNMF}
\title{Evaluate NMF model reconstruction quality}
\usage{
evaluateNMF(x, name = "NMF", assay = "logcounts")
}
\arguments{
\item{x}{A SingleCellExperiment or SpatialExperiment object with NMF results}

\item{name}{Character, name of the NMF result to evaluate (default "NMF")}

\item{assay}{Character, which assay was used for NMF (default "logcounts")}
}
\value{
Numeric scalar, the reconstruction loss (lower is better)
}
\description{
Computes the reconstruction loss for an NMF model against its original
data matrix. Wraps \code{\link[RcppML]{evaluate}} with a
SingleCellExperiment interface.
}
\examples{
library(scuttle)
sce <- mockSCE(ngenes = 100, ncells = 50)
sce <- logNormCounts(sce)
sce <- runNMFscape(sce, k = 5, verbose = FALSE)
evaluateNMF(sce)

}
\seealso{
\code{\link{getModel}} for which functions store an S4
  \code{nmf} model. FactorNet-based results are not supported here;
  use their stored \code{$total_loss} instead.
}
