% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getBasis}
\alias{getBasis}
\title{Extract NMF basis matrix from SingleCellExperiment object}
\usage{
getBasis(x, name = "NMF")
}
\arguments{
\item{x}{A SingleCellExperiment object with NMF results}

\item{name}{Character, name of the NMF result to extract (default "NMF")}
}
\value{
Matrix with features x factors (basis matrix)
}
\description{
Extract NMF basis matrix from SingleCellExperiment object
}
\examples{
library(scuttle)
sce <- mockSCE(ngenes = 100, ncells = 50)
sce <- logNormCounts(sce)
sce <- runNMFscape(sce, k = 3, verbose = FALSE)
basis <- getBasis(sce)
dim(basis)
}
