% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{getTopFeatures}
\alias{getTopFeatures}
\title{Calculate feature loadings (which genes contribute most to each factor)}
\usage{
getTopFeatures(x, name = "NMF", n = 10)
}
\arguments{
\item{x}{A SingleCellExperiment object with NMF results}

\item{name}{Character, name of the NMF result to use (default "NMF")}

\item{n}{Integer, number of top features to return per factor (default 10)}
}
\value{
List of character vectors, each containing top features for a factor
}
\description{
Calculate feature loadings (which genes contribute most to each factor)
}
\examples{
library(scuttle)
sce <- mockSCE(ngenes = 100, ncells = 50)
sce <- logNormCounts(sce)
sce <- runNMFscape(sce, k = 3, verbose = FALSE)
top_genes <- getTopFeatures(sce, n = 10)
head(top_genes[[1]])
}
