% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFscape-package.R
\docType{package}
\name{NMFscape-package}
\alias{NMFscape}
\alias{NMFscape-package}
\title{NMFscape: Fast Non-negative Matrix Factorization for Single Cell and Spatial Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides high-performance non-negative matrix factorization (NMF) methods for SingleCellExperiment and SpatialExperiment objects using the RcppML backend. It enables efficient dimensionality reduction, feature extraction, and pattern discovery in single-cell and spatial genomics data, with seamless integration into Bioconductor workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MicTott/NMFscape}
  \item \url{https://mictott.github.io/NMFscape/}
  \item Report bugs at \url{https://github.com/MicTott/NMFscape/issues}
}

}
\author{
\strong{Maintainer}: Michael Totty \email{mictott@gmail.com} (\href{https://orcid.org/0000-0002-9292-8556}{ORCID})

}
\keyword{internal}
