% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{reconstructNMF}
\alias{reconstructNMF}
\title{Reconstruct original matrix from NMF factors}
\usage{
reconstructNMF(x, name = "NMF")
}
\arguments{
\item{x}{A SingleCellExperiment object with NMF results}

\item{name}{Character, name of the NMF result to use (default "NMF")}
}
\value{
Reconstructed matrix (basis %*% t(coefficients))
}
\description{
Reconstruct original matrix from NMF factors
}
\examples{
library(scuttle)
sce <- mockSCE(ngenes = 100, ncells = 50)
sce <- logNormCounts(sce)
sce <- runNMFscape(sce, k = 3, verbose = FALSE)
reconstructed <- reconstructNMF(sce)
dim(reconstructed)
}
